Dodaj portfel + zakup karnetu

+++ backend/src/main/java/biali/fitmanager/backend/repository/MembershipTypeRepository.java
@@ -3,5 +3,5 @@ package biali.fitmanager.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import biali.fitmanager.backend.model.MembershipType;
 
-public interface MembershipTypeRepository extends JpaRepository<MembershipType, Integer> {
+public interface MembershipTypeRepository extends JpaRepository<MembershipType, Long> {
 }
